moved auth functions into controller directory, WIP: get login and sign up to work
