+++ my-new-project/database/index.js
@@ -10,7 +10,7 @@ const indoorRoutes = require('./routes/indoorRoutes')
 const outdoorRoutes = require('./routes/outdoorRoutes')
 const nightlifeRoutes = require('./routes/nightlifeRoutes')
 const userRoutes = require('./routes/userRoutes')
-const auth = require('./routes/auth')
+const authRoutes = require('./routes/auth')
 
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(bodyParser.json())
@@ -33,7 +33,7 @@ app.use('/indoor', indoorRoutes)
 app.use('/outdoor', outdoorRoutes)
 app.use('/nightlife', nightlifeRoutes)
 app.use('/users', userRoutes)
-app.use('/signup', auth)
+app.use('/auth', authRoutes)
 
 
 app.use(notFound);
